pyproject.toml
[tool.poetry]
name = "corisdb-curation"
version = "0.1.0"
description = ""
authors = ["Benjamin Bearce <bbearce@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.1"
pillow = "^10.3.0"
python-gdcm = "^3.0.23"
pydicom = "^2.4.4"
pypdf2 = "^3.0.1"
psycopg = "^3.1.18"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
couchdb = "^1.2"
exif = "^1.6.0"
google-cloud-bigquery = "^3.20.1"
frontend = "^0.0.3"
opencv-python = "^4.9.0.80"
jupyter = "^1.0.0"
matplotlib = "^3.8.4"
torch = "^2.2.2"
torchvision = "^0.17.2"
ipykernel = "^6.29.4"
pymupdf = "^1.24.11"
bpython = "^0.25"
python-pptx = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
